Refresh of vhost-prefetch-desc-indices

+++ drivers/vhost/vhost.h
@@ -194,7 +194,7 @@ int vhost_get_vq_desc2(struct vhost_virtqueue *,
 void vhost_discard_vq_desc(struct vhost_virtqueue *, int n);
 
 int vhost_prefetch_desc_indices(struct vhost_virtqueue *vq,
-				__virtio16 *indices, int num);
+				__virtio16 *indices, u16 num);
 int vhost_vq_init_access(struct vhost_virtqueue *);
 int vhost_add_used(struct vhost_virtqueue *, unsigned int head, int len);
 int vhost_add_used_n(struct vhost_virtqueue *, struct vring_used_elem *heads,
